pyproject.toml
[project]
name = "sparse-crosscoders-prott5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "requests>=2.32.5",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
